refactor: component structure
