--- packages/gdsc-dju-admin/src/firebase/firebase.config.ts
@@ -1,18 +1,0 @@
-import { initializeApp } from 'firebase/app';
-import { getStorage } from 'firebase/storage';
-
-const firebaseKey = {
-  apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
-  authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
-  databaseURL: import.meta.env.VITE_FIREBASE_DATABASE_URL,
-  projectId: import.meta.env.VITE_FIREBASE_PROJECT_ID,
-  storageBucket: import.meta.env.VITE_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: import.meta.env.VITE_FIREBASE_MESSAGING_SENDER_ID,
-  appId: import.meta.env.VITE_FIREBASE_APPID,
-  measurementId: import.meta.env.VITE_FIREBASE_MEASUREMENT_ID,
-};
-const app = initializeApp(firebaseKey);
-
-const storage = getStorage(app);
-
-export { storage, app };
