+++ packages/gdsc-dju-web/package.json
@@ -11,11 +11,11 @@
     "sitemap": "node sitemap-builder.js"
   },
   "dependencies": {
-    "@gdsc-dju/styled-components": "^0.0.50",
+    "@gdsc-dju/styled-components": "^0.0.53",
     "@gdsc-dju/styled-components-theme": "^0.0.48",
     "axios": "^0.26.1",
     "browser-detect": "^0.2.28",
-    "firebase": "^9.10.0",
+    "firebase": "^9.9.0",
     "firebase-tools": "^11.2.2",
     "framer-motion": "^6.3.0",
     "jotai": "^1.7.7",
@@ -33,8 +33,8 @@
     "swr": "^1.0.1"
   },
   "devDependencies": {
-    "@firebase/app": "^0.7.28",
-    "@firebase/app-compat": "^0.1.29",
+    "@firebase/app": "^0.7.33",
+    "@firebase/app-compat": "^0.1.34",
     "@types/node": "^18.0.4",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
