+++ packages/gdsc-dju-web/src/components/templates/AnnouncementLayout/styled.ts
