+++ packages/gdsc-dju-web/src/pages/Recruit/Recruitment/styled.ts
@@ -4,7 +4,6 @@ import styled from 'styled-components';
 export const CategoryWrapper = styled(motion.div)`
   display: flex;
   padding: 24px 0;
-  margin-right: 30px;
   cursor: pointer;
   border-radius: 30px;
   transition: all 0.3s ease;
@@ -25,30 +24,24 @@ export const CategoryWrapper = styled(motion.div)`
     transition-delay: 0s;
   }
 `;
-export const TeamCategoryWrapper = styled.div`
-  display: flex;
-  padding: 24px 0px;
-  margin-right: 30px;
-  justify-content: space-around;
-  @media (max-width: 500px) {
-    display: none;
-  }
-`;
+
 export const CategoryInner = styled.div`
-  padding-left: 30px;
+  padding-left: 40px;
+  width: 100%;
   font-size: ${({ theme }) => theme.fontSizes.textXl};
-  display: flex;
+  display: grid;
+  grid-template-columns: 1.5fr 1fr 1fr 0.6fr;
   flex-direction: row;
   align-items: center;
   min-width: 320px;
-  @media (max-width: 320px) {
+  @media (max-width: ${({ theme }) => theme.windowSizes.tablet}px) {
     padding-left: 10px;
-    min-width: 150px;
   }
   @media (max-width: 500px) {
     font-size: 18px;
     display: flex;
     flex-direction: column;
+    gap: 2px;
     min-width: 320px;
     align-items: flex-start;
     padding-left: 10px;
@@ -57,54 +50,14 @@ export const CategoryInner = styled.div`
 export const SubCategory = styled.div`
   display: flex;
   align-items: center;
-  width: 300px;
   font-size: 1.3rem;
   color: ${({ theme }) => theme.colors.grey600};
-  @media (max-width: ${({ theme }) => theme.windowSizes.desk}px) {
-    width: 300px;
-  }
-  @media (max-width: ${({ theme }) => theme.windowSizes.tablet}px) {
-    width: 200px;
-  }
-  @media (max-width: ${({ theme }) => theme.windowSizes.mobile}px) {
-    width: 200px;
-  }
 `;
-export const TeamCategory = styled.div`
-  display: flex;
-  align-items: center;
-  margin-left: 30px;
-  width: 330px;
-  font-size: 1.3rem;
-  color: ${({ theme }) => theme.colors.grey700};
-`;
-export const TeamSubCategory = styled.div`
-  display: flex;
-  -webkit-box-align: center;
-  align-items: center;
-  margin-left: 10px;
-  width: 200px;
-  font-size: 1.3rem;
-  color: ${({ theme }) => theme.colors.grey700};
-`;
+
 export const Category = styled.div`
   display: flex;
   font-weight: bold;
   align-items: center;
-  width: 500px;
   font-size: 1.8rem;
-  color: ${({ theme }) => theme.colors.grey800};
-  @media (max-width: ${({ theme }) => theme.windowSizes.desk}px) {
-    width: 500px;
-  }
-  @media (max-width: ${({ theme }) => theme.windowSizes.tablet}px) {
-    width: 300px;
-  }
-  @media (max-width: ${({ theme }) => theme.windowSizes.mobile}px) {
-    width: 200px;
-  }
-  @media (max-width: 320px) {
-    width: 200px;
-    padding: 0;
-  }
+  color: ${({ theme }) => theme.colors.grey900};
 `;
