+++ packages/gdsc-dju-web/src/types/pageData.d.ts
@@ -1,6 +1,17 @@
-export interface RecruitmentDetailType {
-  id: string;
+export type RecruitmentId =
+  | 'frontend'
+  | 'backend'
+  | 'android'
+  | 'design'
+  | 'ml'
+  | 'beginner';
+
+export type RecruitmentStatus = 'OPEN' | 'CLOSED' | 'ALWAYS_OPEN';
+
+export interface AnnouncementData {
+  id: RecruitmentId;
   name: string;
+  status: RecruitmentStatus;
   skill: string;
   stack: string;
   people: string; //이런 분을 찾고 있어요
